Had to rework my review form destination. DestinationsReview is now its own component and its own route in React. I passed props to my existing InfoPage component, and am working on getting these two components to work together.
