+++ front-end-react/doggy-destination-app/src/App.jsx
@@ -6,6 +6,7 @@ import Footer from './components/common/Footer';
 import { BrowserRouter as Router, Route, Routes } from "react-router";
 import SelectedDestination from './components/destinations/SelectedDestination';
 import InfoPage from './components/destinations/InfoPage';
+import DestinationReviews from './components/destinations/DestinationReviews';
 
 
 function App() {
@@ -19,7 +20,8 @@ function App() {
             <Route path="/" element={<Home />} />
             <Route path="/pupPlaces" element={<Main />} />
             <Route path="/idealInfo" element={<SelectedDestination />} />
-            <Route path="/doggyDestinations/:name" element={<InfoPage />} />  
+            <Route path="/doggyDestinations/:name" element={<InfoPage />} />
+            <Route path="/destinationReviews" element={<DestinationReviews />} /> 
           </Routes>
         </Router>        
       </main>
